+++ apps/apigateway/src/core/activity/activity.module.ts
@@ -7,6 +7,7 @@ import { Activity, ActivitySchema } from "./schemas/activity.schema"
 import { ActivityRepository } from "./activity.repository"
 import { CreateActivityCommandHandler } from "./commands/handler/create-activity.handler"
 import { DatabaseModule } from "src/shared/database/database.module"
+import { GetActivityQueryHandler } from "./queries/handler/get-activity-count.handler"
 
 @Module({
   imports: [
@@ -14,7 +15,7 @@ import { DatabaseModule } from "src/shared/database/database.module"
     DatabaseModule.forFeature([{ name: Activity.name, schema: ActivitySchema }], DbConnectionMap.Core),
   ],
   controllers: [ActivityController],
-  providers: [ActivityService, ActivityRepository, CreateActivityCommandHandler],
+  providers: [ActivityService, ActivityRepository, CreateActivityCommandHandler, GetActivityQueryHandler],
 })
 
 export class ActivityModule { }