+++ apps/apigateway/src/core/activity/queries/impl/get-activity-count.query.ts
@@ -1,0 +1,8 @@
+import { GetCountDto } from "../../dto/get-count.dto"
+
+export class GetActivityQuery {
+  constructor(
+    public readonly getCountDto: GetCountDto,
+    public readonly userId: string
+  ) { }
+}