+++ apps/uiservice/src/app/(auth)/dashboard/card/index.tsx
@@ -6,7 +6,7 @@ import {
   CardTitle,
 } from "@/components/ui/card"
 import { brandName, uiConstants } from "@/constants/global-constants"
-import { Circle, Medal, Star } from "lucide-react"
+import { Box, Circle, Medal, Star } from "lucide-react"
 
 interface GenericCardProps {
   productName: string
@@ -15,14 +15,18 @@ interface GenericCardProps {
   category: string
   status: string
   handleClick: (productName: string) => void
+  productIcon: string
 }
 
-export function ProductCard({ productName, status, displayName, desc, category, handleClick }: GenericCardProps) {
+export function ProductCard({ productName, status, displayName, desc, category, handleClick, productIcon }: GenericCardProps) {
   return (
     <Card className="cursor-pointer" onClick={(): void => handleClick(productName)}>
       <CardHeader>
         <div className="space-y-1">
-          <CardTitle className="text-xl">{brandName} {displayName}</CardTitle>
+          <div className="flex justify-between">
+            <CardTitle className="text-xl">{brandName} {displayName}</CardTitle>
+            <div className="scale-75" dangerouslySetInnerHTML={{ __html: productIcon }} />
+          </div>
           <CardDescription>
             {desc}
           </CardDescription>
@@ -31,11 +35,11 @@ export function ProductCard({ productName, status, displayName, desc, category, 
       <CardContent>
         <div className="flex space-x-4 text-sm text-muted-foreground">
           <div className="flex items-center">
-            <Circle className="mr-1 h-3 w-3 fill-sky-400 text-sky-400" />
+            <Box className="mr-1 h-3 w-3 fill-sky-400 text-sky-400" />
             {category}
           </div>
           <div className="flex items-center">
-            <Star className="mr-1 h-3 w-3 fill-yellow-400 text-yellow-400" />
+            <Circle className="mr-1 h-3 w-3 fill-green-400 text-green-400" />
             {status}
           </div>
         </div>